rename bcdb to rpc
